refactor ->  deaggregation module
