+++ setup.cfg
@@ -21,11 +21,13 @@ exclude = .git,
     .pytest_cache,
     .vscode,
     .github,
+    toshi_hazard_store/deaggregate_hazard_mp.py #
     # By default test codes will be linted.
     # tests
 
 [mypy]
 ignore_missing_imports = True
+exclude = toshi_hazard_store/deaggregate_hazard_mp.py
 
 [coverage:run]
 # uncomment the following to omit files during running
