+++ toshi_hazard_store/aggregate_rlzs_mp.py
@@ -9,7 +9,6 @@ from toshi_hazard_store.aggregate_rlzs import (
     concat_df_files,
     get_imts,
     get_levels,
-    process_disagg_location_list,
     process_location_list,
 )
 from toshi_hazard_store.branch_combinator.branch_combinator import (
@@ -23,55 +22,8 @@ from toshi_hazard_store.branch_combinator.SLT_TAG_FINAL import logic_tree_permut
 # from toshi_hazard_store.data_functions import weighted_quantile
 from toshi_hazard_store.locations import locations_nzpt2_chunked
 
-# from toshi_hazard_store.branch_combinator.SLT_TAG_TD import logic_tree_permutations
-# from toshi_hazard_store.branch_combinator.SLT_TAG_TD import data as gtdata
 
-# from toshi_hazard_store.branch_combinator.SLT_TAG_TI import logic_tree_permutations
-# from toshi_hazard_store.branch_combinator.SLT_TAG_TI import data as gtdata
-
-
-class DisaggHardWorker(multiprocessing.Process):
-    """A worker that batches and saves records to DynamoDB.
-
-    based on https://pymotw.com/2/multiprocessing/communication.html example 2.
-    """
-
-    def __init__(self, task_queue, result_queue):
-        multiprocessing.Process.__init__(self)
-        self.task_queue = task_queue
-        self.result_queue = result_queue
-
-    def run(self):
-        print(f"worker {self.name} running.")
-        proc_name = self.name
-
-        while True:
-            nt = self.task_queue.get()
-            if nt is None:
-                # Poison pill means shutdown
-                self.task_queue.task_done()
-                print('%s: Exiting' % proc_name)
-                break
-
-            # tic = time.perf_counter()
-            disagg_configs = process_disagg_location_list(
-                nt.hazard_curves,
-                nt.source_branches,
-                nt.toshi_ids,
-                nt.poes,
-                nt.inv_time,
-                nt.vs30,
-                nt.locs,
-                nt.aggs,
-                nt.imts,
-            )
-            self.task_queue.task_done()
-            self.result_queue.put(disagg_configs)
-
-        return
-
-
-class AggHardWorker(multiprocessing.Process):
+class AggregationWorkerMP(multiprocessing.Process):
     """A worker that batches and saves records to DynamoDB.
 
     based on https://pymotw.com/2/multiprocessing/communication.html example 2.
@@ -108,9 +60,6 @@ class AggHardWorker(multiprocessing.Process):
 
 
 AggTaskArgs = namedtuple("AggTaskArgs", "grid_loc locs toshi_ids source_branches aggs imts levels vs30")
-DisaggTaskArgs = namedtuple(
-    "DisaggTaskArgs", "grid_loc hazard_curves source_branches toshi_ids poes inv_time vs30 locs aggs imts"
-)
 
 
 def build_source_branches(logic_tree_permutations, gtdata, vs30, omit, truncate=None):
@@ -155,7 +104,7 @@ def process_agg(vs30, location_generator, aggs, imts=None, output_prefix='', num
     result_queue: multiprocessing.Queue = multiprocessing.Queue()
 
     print('Creating %d workers' % num_workers)
-    workers = [AggHardWorker(task_queue, result_queue) for i in range(num_workers)]
+    workers = [AggregationWorkerMP(task_queue, result_queue) for i in range(num_workers)]
     for w in workers:
         w.start()
 
@@ -190,69 +139,9 @@ def process_agg(vs30, location_generator, aggs, imts=None, output_prefix='', num
 
     file_name = '_'.join((output_prefix, 'all_aggregates.json'))
     hazard_curves = concat_df_files(df_file_names)
-
     hazard_curves.to_json(file_name)
 
     return hazard_curves, source_branches
-
-    # hazard_curves = pd.concat([hazard_curves, binned_hazard_curves])
-    # print(hazard_curves)
-
-
-def process_disaggs(
-    hazard_curves, source_branches, poes, inv_time, vs30, location_generator, aggs, imts, num_workers=12
-):
-
-    # write serial code for now, parallelize once it works
-    omit = []
-    toshi_ids = [b.hazard_solution_id for b in merge_ltbs_fromLT(logic_tree_permutations, gtdata=gtdata, omit=omit)]
-
-    task_queue: multiprocessing.JoinableQueue = multiprocessing.JoinableQueue()
-    result_queue: multiprocessing.Queue = multiprocessing.Queue()
-
-    print('Starting Disaggregations')
-    print('Creating %d workers' % num_workers)
-    workers = [DisaggHardWorker(task_queue, result_queue) for i in range(num_workers)]
-    for w in workers:
-        w.start()
-
-    tic = time.perf_counter()
-    # Enqueue jobs
-    num_jobs = 0
-    for key, locs in location_generator().items():
-        print(locs)
-        t = DisaggTaskArgs(key, hazard_curves, source_branches, toshi_ids, poes, inv_time, vs30, locs, aggs, imts)
-        task_queue.put(t)
-        num_jobs += 1
-
-    # Add a poison pill for each to signal we've done everything
-    for i in range(num_workers):
-        task_queue.put(None)
-
-    # Wait for all of the tasks to finish
-    task_queue.join()
-
-    # Start printing results
-    print('Results:')
-    disagg_configs = []
-    while num_jobs:
-        result = result_queue.get()
-        disagg_configs += result
-        print(str(result))
-        num_jobs -= 1
-
-    toc = time.perf_counter()
-    print(f'time to run disaggregations: {toc-tic:.0f} seconds')
-
-    return disagg_configs
-
-    # ========================================#
-
-    # disagg_configs = []
-    # for key, locs in location_generator().items():
-    #     disagg_configs += process_disagg_location_list(hazard_curves, source_branches, toshi_ids,
-    #         poes, inv_time, vs30, locs, aggs, imts)
-    # return disagg_configs
 
 
 if __name__ == "__main__":
@@ -262,7 +151,6 @@ if __name__ == "__main__":
     logging.getLogger('toshi_hazard_store').setLevel(logging.DEBUG)
 
     nproc = 20
-
     classical = True
 
     vs30 = 400
@@ -285,53 +173,12 @@ if __name__ == "__main__":
 
     output_prefix = f'FullLT_{loc_keyrange[0]}_{loc_keyrange[1]}'
 
-    if classical:
-        hazard_curves, source_branches = process_agg(
-            vs30,
-            location_generator,
-            aggs,
-            imts,
-            output_prefix=output_prefix,
-            num_workers=nproc,
-            location_range=loc_keyrange,
-        )
-
-    # # if running classical and disagg you must make sure that the requested locations, imts, vs30, aggs for disaggs
-    # # are in what was requested for the classical calculation
-    # disaggs = False
-    # poes = [0.1, 0.02]
-    # aggs = ['mean']
-    # inv_time = 50
-    # imts = ['PGA', 'SA(0.5)', 'SA(1.5)']
-    # output_prefix = 'fullLT_dissags'
-    # location_generator = locations_nz34_chunked
-    # # location_generator = locations_nz2_chunked
-    # # breakpoint()
-
-    # if disaggs:
-    #     if classical:
-    #         disagg_configs = process_disaggs(
-    #             hazard_curves, source_branches, vs30, location_generator, aggs, imts, num_workers=nproc
-    #         )
-    #     else:
-    #         hazard_curves, source_branches = process_agg(
-    #             vs30, location_generator, aggs, imts, output_prefix='for_disaggs', num_workers=nproc
-    #         )
-    #         disagg_configs = process_disaggs(
-    #             hazard_curves, source_branches, poes, inv_time, vs30, location_generator, aggs,
-    #                imts, num_workers=nproc
-    #         )
-
-    #     # add location code for sites that have them
-    #     from nzshm_common.location.code_location import CodedLocation
-    #     from nzshm_common.location.location import LOCATIONS_BY_ID
-
-    #     for disagg_config in disagg_configs:
-    #         for loc in LOCATIONS_BY_ID.values():
-    #             location = CodedLocation(loc['latitude'], loc['longitude']).downsample(0.001).code
-    #             if location == disagg_config['location']:
-    #                 disagg_config['site_code'] = loc['id']
-    #                 disagg_config['site_name'] = loc['name']
-
-    #     with open('disagg_configs.json', 'w') as json_file:
-    #         json.dump(disagg_configs, json_file)
+    hazard_curves, source_branches = process_agg(
+        vs30,
+        location_generator,
+        aggs,
+        imts,
+        output_prefix=output_prefix,
+        num_workers=nproc,
+        location_range=loc_keyrange,
+    )
